T-04 - Add nibbles banner to home page
Summary:
This commit adds the nibbles banner to the home page.

Folder layout structure was added for assets/images,
components, pages, and layouts as there was no folder
structure previously available.

CSS is used instead of JSX styles as regular CSS is easier
to use with media queries.

A layout component was made with responsiveness so that
the default webpage layout can be reused elsewhere.

No react routing has been added as this should be done
in a separate commit.

No burger menu is used as this could be a future feature
and is currently not the priority. The page does
respond from 320px upwards though.