E -04-U-01 Ability to use navbar to traverse the site
Summary:
This commit adds react router navigation to the navbar so users can
go to different pages. Additionally, login and signup modal templates
were added with working buttons to close the modal were added to the
header component. These modals are a part of navigation in the eyes
of the user. Some minor refactor and improvements were added to the
header and buttons to make them more reusable.