E-05-U-01-T-01 Create recipe thumbnail component
Summary:
This commit adds a react component to show a thumbnail of a recipe from
a search result. Modals will be added in the future to navigate and show
additional images as well as show the ingredients list. This will be done
in the future once images urls are finalised in a known data format
and also the list of ingredients can be passed down.